Changed ClassType and MetaclassType to work correctly with metaclasses. Set up correct hierarchies for classes, metaclasses and metaclass classes. Tidied up some code in ClassType. Established a TODO list on the Argon struct to log things that need to be down.  Created a TODO list for logging semantic checks that need to be done before or after type inference.

+++ Argon/Main UI/Project Editor/Hierarchy/IDE Nodes/IDENode.swift
@@ -136,7 +136,7 @@ public class IDENode: NSObject,NSCoding,Comparable,Dependent
         self.name = coder.decodeObject(forKey: "name") as! String
         self.path = coder.decodePath(forKey: "path")!
         self.nodeKey = coder.decodeInteger(forKey: "nodeKey")
-        self.sourceItemState = IDESourceItemState(rawValue: UInt8(coder.decodeInteger(forKey: "sourceItemState")))
+        self.sourceItemState = IDESourceItemState(rawValue: coder.decodeInteger(forKey: "sourceItemState"))
         }
         
     public func encode(with coder: NSCoder)
